requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pymysql
mysql-connector-python
python-dotenv

#-e . # This line allows you to install the current directory as a package, if you have a setup.py file.